+++ changelog.md
@@ -2,3 +2,4 @@
 - Improved meeting point generation
 - Fixed trimmed planks item tag being incomplete
 - Fixed parts of the inverted bell being broken by water
+- Fixed several bugs with the tortoise and improved their performance